rename "Enhancers" dir -> "enhancers"
