+++ src/containers/QuestionnaireDesignPage/index.js
@@ -2,9 +2,9 @@ import { compose } from "react-apollo";
 import { connect } from "react-redux";
 
 import QuestionnaireDesign from "./QuestionnaireDesignPage";
-import withQuestionnaire from "../Enhancers/withQuestionnaire";
-import withUpdateSection from "../Enhancers/withUpdateSection";
-import withUpdatePage from "../Enhancers/withUpdatePage";
+import withQuestionnaire from "../enhancers/withQuestionnaire";
+import withUpdateSection from "../enhancers/withUpdateSection";
+import withUpdatePage from "../enhancers/withUpdatePage";
 import getUrlParams from "../../utils/getUrlParams";
 
 export const mapStateToProps = (state, { match }) => getUrlParams(match.params);
