+++ src/containers/enhancers/withDeleteAnswer.js
