+++ src/containers/enhancers/withDeleteOption.js
