+++ src/containers/enhancers/withDeleteOption.test.js
