+++ src/containers/enhancers/withDeletePage.js
