+++ src/containers/enhancers/withDeletePage.test.js
